+++ inst/NEWS.md
@@ -6,5 +6,5 @@
 
 ## **Version 0.1-2**
 
-- New functions :  vip, orthog, sopls, soplsrda, soplslds, soplsqda
+- New functions :  vip, orthog, mbplsr, mbplskern, mbplsrannar, mbplsnipals, mbplsrda, mbplslda, mbplsqda, soplsr, soplsrda, soplslda, soplsqda
 - Function changes : "scaling" argument added to plskern, plsrannar, plsnipals, plsrda, plslda, plsqda functions