+++ inst/NEWS.md
@@ -3,3 +3,10 @@
 ## **Version 0.1-1**
 
 - Submission to CRAN.
+
+## **Version 0.1-2**
+
+- New functions :  vip, orthog, mbplsr, mbplsrda, mbplslda, mbplsqda, soplsr, soplsrda, soplslda, soplsqda, soplsrcv, soplsrdacv, soplsldacv, soplsqdacv
+- Function changes : "Xscaling" and "Yscaling" arguments added to plskern, plsrannar, plsnipals, plsrda, plslda, plsqda functions
+- bugfix R/plotjit.R; R/plotscore.R; R/plotxy.R; R/selwold.R.
+